Correct formatting according to codeclimate suggestions
